requirements.txt
Django==4.2.1
django_environ==0.10.0
django_tinymce==3.6.1
razorpay==1.3.0
Pillow==9.5.0
psycopg2-binary==2.9.6